requirements.txt
einops==0.6.1
matplotlib==3.7.1
numpy==1.24.2
opencv_python==4.7.0.72
Pillow==9.3.0
scikit_learn==1.2.2
torchvision==0.15.1
gdown